tr1d1um: add client package
